fix for double to float warning in MSVC
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@2205 d7cf8633-e32d-0410-b094-e92efae38249

+++ cgame.c
@@ -48,7 +48,7 @@ static localentity_t *entspawn(void)
 {
 	int i, best;
 	float bestfreetime;
-	bestfreetime = gametime + 100.0f;
+	bestfreetime = (float) (gametime + 100.0);
 	best = -1;
 	for (i = 0;i < MAX_LOCALENTITIES;i++)
 	{
